feat: updates deletePost logic
Removes the unnecessary logic of injecting the GetPostById in the DeletePostController, and do all the logic only in the handler.

+++ VSAProject.API/Features/DeletePost/DeletePostHandler.cs
@@ -6,11 +6,16 @@ namespace VSAProject.API.Features.DeletePost
     {
         private readonly AppDbContext _db = db;
 
-        public async Task Handle(int id)
+        public async Task<bool> Handle(int id)
         {
             var post = _db.Posts.First(p => p.Id == id);
+            if (post == null)
+            {
+                return false;
+            }
             _db.Remove(post);
             await _db.SaveChangesAsync();
+            return true;
         }
     }
 }
